refactor(export): keep animation recovery in HTML capture
